% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_df_report_no_rownames}
\alias{add_df_report_no_rownames}
\title{Add a dataframe to a report.}
\usage{
add_df_report_no_rownames(report, csv_relative_path)
}
\arguments{
\item{report}{Path to the report.}

\item{csv_relative_path}{Path to the csv, relative to the report.}
}
\value{
Nothing
}
\description{
Will read a csv file and display its content on the report.
}
