% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_image_markdown_report}
\alias{add_image_markdown_report}
\title{Add an image to a report}
\usage{
add_image_markdown_report(report, image_relative_path, title = "")
}
\arguments{
\item{report}{Path to the report.}

\item{image_relative_path}{Path to the image, relative to the report.}

\item{title}{Title to give the image}
}
\value{
Nothing
}
\description{
Add an image to a report
}
