% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_text_report}
\alias{add_text_report}
\title{Add some text to a report}
\usage{
add_text_report(report, txt = "")
}
\arguments{
\item{report}{Path to the report.}

\item{txt}{Text to add. Default to ""}
}
\value{
Nothing
}
\description{
Add some text to a report
}
