% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_textfile_report}
\alias{add_textfile_report}
\title{Add a chunk of code to read content of file in a report.}
\usage{
add_textfile_report(report, relative_textfile)
}
\arguments{
\item{report}{Path to the report.}

\item{relative_textfile}{Path to the text file, relative to the report.}
}
\value{
Nothing
}
\description{
Add a chunk of code to read content of file in a report.
}
