% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{add_title_report}
\alias{add_title_report}
\title{Add a markdown title to a report}
\usage{
add_title_report(report, title = "", level = 1)
}
\arguments{
\item{report}{Path to the report.}

\item{title}{Content of the title. Default to ""}

\item{level}{A number between 1 and 5, 1 being a title and going up means a lower level of subtitle. Default to 1}
}
\value{
Nothing
}
\description{
Add a markdown title to a report
}
