% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{build_report}
\alias{build_report}
\title{Render the markdown document.}
\usage{
build_report(report, title)
}
\arguments{
\item{report}{Path to the report.}

\item{title}{Title to give to the report}
}
\value{
Nothing
}
\description{
Render the markdown document.
}
