% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{CellQc}
\alias{CellQc}
\alias{CellQc,Seurat-method}
\title{Cells in the quality control}
\usage{
CellQc(object, ...)

\S4method{CellQc}{Seurat}(
  object,
  min.nFeature_RNA = 200,
  max.nFeature_RNA = Inf,
  max.percent.mt = 5,
  ...
)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{min.nFeature_RNA}{Include cells where at least this many features are detected, default 200}

\item{max.nFeature_RNA}{Include cells where limits this many features are detected, defalult Inf}

\item{max.percent.mt}{Include cells where largest proportion of mitochondrial genes, default five percent}
}
\value{
Returns the filtered Seruat object.
}
\description{
Cells in the quality control
}
