% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{check_assay}
\alias{check_assay}
\title{Check that assay is in seurat}
\usage{
check_assay(seurat, assay = "RNA")
}
\arguments{
\item{seurat}{an object to check}

\item{assay}{an assay to check}
}
\value{
nothing
}
\description{
Check if object is of class seurat, check that assay is a string, and check that assay is present in seurat
}
