% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chunked.read}
\alias{chunked.read}
\title{random slicing by reading data by chunks (entire columns)}
\usage{
chunked.read(x, idx, verbose = FALSE, mc.cores = 1, fast = FALSE, ...)
}
\arguments{
\item{x}{DelayedArray seed object that supports 'dim' and 'chunk_selection'}

\item{idx}{row and col index}
}
\description{
random slicing by reading data by chunks (entire columns)
}
