% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{clustering}
\alias{clustering}
\title{Perform the clustering step, and optionally a clustree graph}
\usage{
clustering(
  analysis,
  sample = "",
  method = "Seurat",
  res_clustree = c(),
  resolution = 1,
  plots_dir = ""
)
}
\arguments{
\item{analysis}{The analysis object.}

\item{sample}{Sample name, will be used to name the images files.}

\item{method}{The analysis method. Default "Seurat"}

\item{res_clustree}{Vector of resolutions to use in the clustree.
Typically resolutions range between 0.1 and 2. Keep empty to skip. Default c()}

\item{resolution}{Which final resolution to keep (can be a resolution not in \code{res_clustree}).
Typically resolutions range between 0.1 and 2. Default 1}

\item{plots_dir}{The path to save the plots. Keep empty to skip plot saving. Default ""}
}
\value{
An analysis object of type method with clusters.
}
\description{
Perform the clustering step, and optionally a clustree graph
}
