% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_tube.R
\name{compute_curve_sd}
\alias{compute_curve_sd}
\title{Compute the bootstrap uncertainty}
\usage{
compute_curve_sd(target_curve_list, bootstrap_curve_list)
}
\arguments{
\item{target_curve_list}{a list}

\item{bootstrap_curve_list}{a list}
}
\value{
a numeric
}
\description{
Compute the bootstrap uncertainty
}
