% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_3dplot.R
\name{construct_3d_tube}
\alias{construct_3d_tube}
\title{Construct the 3D tube for each curve}
\usage{
construct_3d_tube(dat, radius, len = 20)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{radius}{a numeric}

\item{len}{a numeric}
}
\value{
a list
}
\description{
Construct the 3D tube for each curve
}
