% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{CreateSingleRNA}
\alias{CreateSingleRNA}
\title{Create DataSheet}
\usage{
CreateSingleRNA(data.path, sample.name)
}
\arguments{
\item{data.path}{Vector of data path}

\item{sample.name}{Vector of data name}
}
\value{
DataSheet object
}
\description{
Use input data information create a DataSheet object
}
\examples{
data <- CreateSingleRNA(data.path = c("./data/filtered_gene_bc_matrices3k/hg19/", "./data/filtered_feature_bc_matrix10k/"),
                        sample.name = c("pbmc", "lymphoma"))

}
