% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\docType{class}
\name{DataSheet-class}
\alias{DataSheet-class}
\alias{DataSheet}
\title{The DataSheet Class}
\description{
The DataSheet class is an intermediate data storage class that sotres the data path
}
\section{Slots}{

\describe{
\item{\code{data.path}}{Vector of data path}

\item{\code{sample.name}}{Vector of data name}
}}

