% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{diagnosis}
\alias{diagnosis}
\title{Title}
\usage{
diagnosis(
  analysis,
  save_path = ".",
  file_name = "analysis",
  organism = "human",
  mitochondrial_genes = c(),
  force_report = FALSE
)
}
\arguments{
\item{analysis}{}

\item{save_path}{}

\item{file_name}{}

\item{organism}{}

\item{mitochondrial_genes}{}

\item{force_report}{}
}
\description{
Title
}
