% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.compute_cluster_center}
\alias{.compute_cluster_center}
\title{Compute the cluster centers}
\usage{
.compute_cluster_center(dat, cluster_mat)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_mat}{a 0-1 matrix that is \code{n} by \code{k}}
}
\value{
a \code{k} by \code{d} matrix
}
\description{
Compute the cluster centers
}
