% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.compute_ellipse_points}
\alias{.compute_ellipse_points}
\title{Compute the points of the ellipse on a two-dimensional plot}
\usage{
.compute_ellipse_points(mean_vec, cov_mat, scale_factor = 1)
}
\arguments{
\item{mean_vec}{2-dimensional vector}

\item{cov_mat}{2 by 2 PSD matrix}

\item{scale_factor}{scaling factor to determine the size of ellipse}
}
\value{
matrix
}
\description{
Function is for plotting purposes only
}
