% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.construct_average_curve}
\alias{.construct_average_curve}
\title{Construct average curve}
\usage{
.construct_average_curve(pcurve_list, dat)
}
\arguments{
\item{pcurve_list}{list of outputs of \code{princurve::project_to_curve()}}

\item{dat}{a \code{n} by \code{d} matrix}
}
\value{
a \code{principal_curve} object
}
\description{
Construct average curve
}
