% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.construct_knn_graph}
\alias{.construct_knn_graph}
\title{Construct the K-nearest neighbor graph based on Euclidean distance}
\usage{
.construct_knn_graph(dat, knn = 5)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{knn}{positive integer}
}
\value{
\code{igraph} graph object
}
\description{
Construct the K-nearest neighbor graph based on Euclidean distance
}
