% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.construct_lineages}
\alias{.construct_lineages}
\title{Construct the lineages}
\usage{
.construct_lineages(spt_graph, starting_cluster)
}
\arguments{
\item{spt_graph}{\code{igraph} object}

\item{starting_cluster}{positive integer}
}
\value{
a list
}
\description{
Enumerates the shortest paths from \code{starting_cluster} to
each of the leaves in \code{spt_graph}
}
