% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.construct_spt}
\alias{.construct_spt}
\title{Construct the shortest path tree graph from KNN graph}
\usage{
.construct_spt(knn_graph, k, starting_cluster)
}
\arguments{
\item{knn_graph}{\code{igraph} object}

\item{k}{positive integer for number of clusters}

\item{starting_cluster}{the "origin" cluster that all the lineages will start
from}
}
\value{
\code{igraph} object representing the shortest path tree
}
\description{
This is currently hard-coded to work for only one \code{starting_cluster}.
}
\details{
Note: The squaring of \code{dist_mat} is arbitrary, currently used to
encourage paths through other clusters.
}
