% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.determine_idx_lineage}
\alias{.determine_idx_lineage}
\title{Determine which samples are part of a lineage}
\usage{
.determine_idx_lineage(lineage, cluster_mat)
}
\arguments{
\item{lineage}{one specific element of the list output of \code{.get_lineages()}}

\item{cluster_mat}{0-1 matrix output of \code{.construct_cluster_matrix()} that
has \code{n} rows and \code{k} column}
}
\description{
Determine which samples are part of a lineage
}
