% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.get_curves}
\alias{.get_curves}
\title{Estimate the slingshot curves}
\usage{
.get_curves(dat, cluster_labels, lineages, shrink = 1, thresh = 0.001,
  max_iter = 15, b = 1)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}}

\item{lineages}{output of \code{.get_lineage()}}

\item{shrink}{shrinkage factor}

\item{thresh}{parameter to determine convergence}

\item{max_iter}{maximum number of iterations}

\item{b}{parameter for the kernel function (when smoothing)}
}
\value{
a list of \code{principal_curve} objects
}
\description{
Estimate the slingshot curves
}
