% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_curves.R
\name{.initialize_curve_hierarchy}
\alias{.initialize_curve_hierarchy}
\title{Constructs the order to average the lineages}
\usage{
.initialize_curve_hierarchy(lineages, cluster_vec)
}
\arguments{
\item{lineages}{output of \code{.get_lineages()}}

\item{cluster_vec}{vector of cluster labels}
}
\value{
list of lineages to average
}
\description{
As stated in the paper, "Average curves are constructed  in a recursive
manner, from the latest branching events to the earliest."
}
