% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_factorization.R
\name{.projection_l1}
\alias{.projection_l1}
\title{L1 projection of the product of two vectors to become nonpositive}
\usage{
.projection_l1(current_vec, other_mat, idx = 1:nrow(other_mat),
  tol = 1e-06)
}
\arguments{
\item{current_vec}{vector}

\item{other_mat}{matrix}

\item{idx}{row indices for other_mat}

\item{tol}{numeric}
}
\description{
Solves the linear program:
Data: u_i (vector), i from 1 to k. V (matrix), with d rows and k columns.
Variables: s_1 to s_k, u_+1 to u_+k, and u_-1 to u_-k.
Optimization problem: min s_1 + ... + s_k
such that: s_i >= u_i - (u_+i - u_-i) for i from 1 to k,
           s_i >= -(u_i - (u_+i - u_-i)) for i from 1 to k,
           V %*% (u_+(1:k) - u_-(1:k)) <= 0 elementwise (for entire vector of length d)
           s_i, u_+i, u_-i >= 0 for i from 1 to k
}
