% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.remove_outliers}
\alias{.remove_outliers}
\title{Remove outliers based on shortest path centrality}
\usage{
.remove_outliers(graph, cluster_labels, percentage = 0.05)
}
\arguments{
\item{graph}{\code{igraph} object}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels, na.rm = T)}. Can include \code{NA}}

\item{percentage}{percentage of vertices to remove that have
a cluster label}
}
\value{
indices (of the original dataset, reflected by \code{cluster_labels})
that should be kept
}
\description{
This function relies on the specific format of \code{graph}.
Specifically, the first \code{max(cluster_labels, na.rm=T)} nodes
in \code{graph} correspond to the clutser centers and are not
reflected with \code{cluster_labels}
}
