% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downsample.R
\name{downsample}
\alias{downsample}
\title{Downsample a dataset}
\usage{
downsample(dat, downsample_rate = 0.5, dropoff_rate = 0.2)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{downsample_rate}{numeric of the rate of downsample (higher means less cells are kept)}

\item{dropoff_rate}{numeric of the rate of dropoff (higher means less cells are kept)}
}
\value{
list containing \code{dat}, \code{downsample_idx}, and \code{dropout_idx}
}
\description{
Downsample a dataset
}
