% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{dropout}
\alias{dropout}
\title{Determine which entries are candidates for dropout}
\usage{
dropout(dat)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}
}
\value{
a 0-1 matrix where 0 indicates entries that are dropout candidates
}
\description{
Determine which entries are candidates for dropout
}
