% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{dropout_create}
\alias{dropout_create}
\title{Create the dropout function}
\usage{
dropout_create(a = 0.7, b = 1)
}
\arguments{
\item{a}{intercept of logistic function}

\item{b}{slope of logistic function}
}
\value{
function
}
\description{
Create the dropout function
}
