% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_estimation.R
\name{em_mixture}
\alias{em_mixture}
\title{EM mixture for modeling dropout}
\usage{
em_mixture(x, mixture = "gamma.tgaussian", min_val = 0,
  max_iter = 100)
}
\arguments{
\item{x}{vector}

\item{mixture}{family, represented by a string}

\item{min_val}{minimum value used for initialization}

\item{max_iter}{maximum iteration for EM algorithm}
}
\value{
parameter object
}
\description{
EM mixture for modeling dropout
}
