% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropout.R
\name{estimate_dropout}
\alias{estimate_dropout}
\title{Estimate dropout function}
\usage{
estimate_dropout(dat, k = 5, lambda = 0.01, threshold_quant_degree = 0.75,
  threshold_quant_logistic = 0.5)
}
\arguments{
\item{dat}{numeric matrix with 0's}

\item{k}{rank of prediction for \code{.funk_svd_prediction}}

\item{lambda}{tuning parameter for \code{.funk_svd_prediction}}

\item{threshold_quant_degree}{parameter for \code{.largest_average_degree_subgraph}}

\item{threshold_quant_logistic}{parameter for \code{.logistic_regression}}
}
\value{
numeric vector
}
\description{
The dropout function is a logistic function, where this function estimates
the intercept and slope.
}
