% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{filter_data}
\alias{filter_data}
\title{Filter out genes and cells outside the min and max limits for several parameters.}
\usage{
filter_data(
  analysis,
  sample,
  assay = "RNA",
  method = "Seurat",
  organism = "",
  mitochondrial_genes = c(),
  min_genes = 100,
  min_counts = 0,
  min_cells = 1,
  min_mt = 0,
  max_genes = Inf,
  max_counts = Inf,
  max_cells = Inf,
  max_mt = 20,
  plots_dir = ""
)
}
\arguments{
\item{analysis}{The analysis object.}

\item{sample}{Sample name, will be used to save the images.}

\item{method}{The analysis method. Default Seurat}

\item{organism}{An organism supported by the library for automatic mitochondrial detection. Currently supported : human and mouse.
Default "" to skip automatic mitochondrial detection.}

\item{mitochondrial_genes}{A vector of mitochondrial genes. If a supported organism is requested, it will be skipped. Default empty}

\item{min_genes}{All cells having a lower number of genes expressed will be filtered out. Default 0}

\item{min_counts}{All cells having a lower number of counts will be filtered out. Default 0}

\item{min_cells}{All genes having a lower number of cells expressing it will be filtered out. Default 0}

\item{min_mt}{All cells having a lower percentage of mitochondria will be filtered out. Default 0}

\item{max_genes}{All cells having a higher number of genes expressed will be filtered out. Default Inf}

\item{max_counts}{All cells having a higher number of counts will be filtered out. Default Inf}

\item{max_cells}{All genes having a higher number of cells expressing it will be filtered out. Default Inf}

\item{max_mt}{All cells having a higher percentage of mitochondria will be filtered out. Default Inf}

\item{plots_dir}{The path to save the plots. Keep empty to skip plot saving. Default ""}
}
\value{
An object of the analysis type filtered.
}
\description{
Filter out genes and cells outside the min and max limits for several parameters.
}
