% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{find_all_DE}
\alias{find_all_DE}
\title{Find DE between all clusters}
\usage{
find_all_DE(
  analysis,
  sample = "",
  method = "Seurat",
  test = "wilcox",
  logfc_threshold = 0.25,
  pvalue_threshold = 0.05,
  results_dir = "",
  variable = "seurat_clusters",
  force_DE = FALSE
)
}
\arguments{
\item{analysis}{The analysis object.}

\item{sample}{Sample name, will be used to name the table files.}

\item{method}{The analysis method. Default "Seurat"}

\item{test}{Which statistical tests to use for DE. To see available options, see the documentation for Seurat::FindAllMarkers. Default "wilcox"}

\item{logfc_threshold}{Filter genes based on a minimum logged fold change. Default 0.25}

\item{pvalue_threshold}{Filter genes based on a maximum pvalue. Default 0.05}

\item{results_dir}{The path to save the tables. Keep empty to skip plot saving. Default ""}

\item{variable}{what to isolate from the FetchData. Default "seurat_clusters"}
}
\value{
A dataframe that contains the differentially expressed genes for each cluster against all other clusters.
}
\description{
Find DE between all clusters
}
