% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_zeros.R
\name{find_true_zeros}
\alias{find_true_zeros}
\title{Finding the true zeros}
\usage{
find_true_zeros(dropout_mat, num_neighbors = NA)
}
\arguments{
\item{dropout_mat}{a \code{n} by \code{d} matrix}

\item{num_neighbors}{number of neighbors}
}
\value{
a 0-1-NA matrix of size \code{n} by \code{d}
}
\description{
Finding the true zeros
}
