% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{get_demo_seurat_object}
\alias{get_demo_seurat_object}
\title{Get demo seurat object}
\usage{
get_demo_seurat_object()
}
\value{
Seurat object
}
\description{
Get demo seurat object
}
\examples{
pbmc <- get_demo_seurat_object()

}
