% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_IO.R
\name{h5read.chunked.character}
\alias{h5read.chunked.character}
\title{the API only reads data from H5 by chunks (entire columns)}
\usage{
\method{h5read.chunked}{character}(x, name, index, ncol, verbose = FALSE,
  mc.cores = 1, fast = TRUE, ...)
}
\description{
the API only reads data from H5 by chunks (entire columns)
}
