% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_IO.R
\name{h5read.chunked}
\alias{h5read.chunked}
\title{the API only reads data from H5 by chunks (entire columns)}
\usage{
h5read.chunked(h5, name, index, nGenes)
}
\description{
the API only reads data from H5 by chunks (entire columns)
}
