% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_IO.R
\name{H5transpose}
\alias{H5transpose}
\title{transpose a 2d data mattrix in H5}
\usage{
H5transpose(src, dest, col.ind = NULL, compress = c("lz4", "gzip", "none"),
  nLevel = NULL, verbose = FALSE, ...)
}
\arguments{
\item{src}{the H5 file that contains the 'data' mattrix to be transposed}

\item{dest}{the H5 file name to be written to}
}
\description{
transpose a 2d data mattrix in H5
}
