% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/H5_IO.R
\name{H5write.blocks}
\alias{H5write.blocks}
\title{Create the H5 file from matrix}
\usage{
H5write.blocks(mat, h5file, ncol, nrow, compress = c("lz4", "gzip", "none"),
  nLevel = NULL)
}
\description{
Create the H5 file from matrix
}
