% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{initialize_report}
\alias{initialize_report}
\title{Initialize a Rmd report from the predefined template.}
\usage{
initialize_report(name, report_path = ".", force = FALSE)
}
\arguments{
\item{name}{Name of the file}

\item{report_path}{Path to where the report should be located. Must be the upper directory of the plots directory (if /path/to/plot is the plot directory, then report should be in /path/to). Default to "."}

\item{force}{Set to true if you want to overwrite the the report that could already exists\code{report_path}. Default FALSE}
}
\value{
Nothing
}
\description{
Initialize a Rmd report from the predefined template.
}
