% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{initialize_report_qmd}
\alias{initialize_report_qmd}
\title{Title}
\usage{
initialize_report_qmd(name, report_path = ".", force = FALSE)
}
\arguments{
\item{name}{}

\item{force}{}

\item{report_part}{}
}
\description{
Title
}
