% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{integrate}
\alias{integrate}
\title{Integrate data from several samples.}
\usage{
integrate(
  samples,
  path_to_count = "",
  analysis_list = list(),
  step = "loading_data",
  from = "Cellranger",
  method = "Seurat",
  assay = "RNA",
  save_path = ".",
  file_name = "integrated",
  make_report = TRUE,
  organism = "",
  mitochondrial_genes = c(),
  min_genes = 0,
  min_counts = 0,
  min_cells = 1,
  min_mt = 0,
  max_genes = Inf,
  max_counts = Inf,
  max_cells = Inf,
  max_mt = Inf,
  nfeatures_normalize = 2000,
  selection_method_normalize = "vst",
  npcs_pca = 50,
  nfeatures_integration = 5000,
  force_report = FALSE,
  k.weight = 100,
  k.filter = 100,
  perform_clusterisation = TRUE
)
}
\arguments{
\item{samples}{A list of the sample names, used for matrix detection and plots saving. This parameter is mandatory}

\item{path_to_count}{Function need either that if step is 1 or \code{analysis_list} if step is 2 to 5.
The path to where the directories for each sample are. The function will automatically
detect the structure in each directory to find the matrix, using the \code{from} parameter. Default ""}

\item{analysis_list}{Function need either that if step is 2 to 5 or \code{analysis_list} if step is 1.
A list of all the objects used for the integration. Default list()}

\item{step}{Steps ranging from \code{loading_data} to \code{integrating}. The following steps are : \code{loading_data}, \code{filtering}, \code{normalizing}, \code{PCA}, and \code{integrating}.
Choosing \code{loading_data} will do the entire process, while \code{integrating} will only do the integration and assume that every previous step
needed has already been done by the user. Default 1}

\item{from}{From which tools the count matrix comes from. Currently supports : Cellranger. Default "Cellranger"}

\item{method}{The analysis method used through the analysis, as well as the objects type in \code{analysis_list}.
Currently supports : Seurat. Default "Seurat"}

\item{assay}{Which to use for all functions asking for it. Default "RNA"}

\item{save_path}{In which directory the plots, results and report are saved. Empty to not save anything. Default "."
Keep empty to just return the analysis and not save it. Default "."}

\item{file_name}{What name to give to the final file. Will be skipped if \code{save_path} is empty. Default "analysis.rds"}

\item{make_report}{Whether to make a report with the plot generated or not.
If \code{save_path} is empty, this parameter will be skipped. Default TRUE}

\item{organism}{The organism is used for mitochondrial genes detection. Supports mouse and human.
Keep empty to skip mitochondrial detection. Default ""}

\item{mitochondrial_genes}{List of mitochondrial genes. This argument will be completely ignored
if a supported organism is provided. Leave \code{organism} and \code{mitochondrial_genes} empty to skip mitochondrial filtering.}

\item{min_genes}{All cells having a lower number of genes expressed will be filtered out. Default 0}

\item{min_counts}{All cells having a lower number of counts will be filtered out. Default 0}

\item{min_cells}{All genes having a lower number of cells expressing it will be filtered out. Default 0}

\item{min_mt}{All cells having a lower percentage of mitochondria will be filtered out. Default 0}

\item{max_genes}{All cells having a higher number of genes expressed will be filtered out. Default Inf}

\item{max_counts}{All cells having a higher number of counts will be filtered out. Default Inf}

\item{max_cells}{All genes having a higher number of cells expressing it will be filtered out. Default Inf}

\item{max_mt}{All cells having a higher percentage of mitochondria will be filtered out. Default Inf}

\item{nfeatures_normalize}{Number of variable genes to select for scaling. Default 2000}

\item{selection_method_normalize}{What method to use for features selection during scaling. Default "vst"}

\item{npcs_pca}{Number of principal components to compute in the PCA. Default 50}

\item{nfeatures_integration}{Number of features to use in the integration process. Default 5000}

\item{force_report}{Whether to overwrite the report if it already exists. If this is FALSE, and a file \code{integration.Rmd} already exists,
report creation will be skipped. Default FALSE}

\item{k.weight}{Number of neighbors to consider when weighting anchors}

\item{k.filter}{How many neighbors (k) to use when filtering anchors}

\item{perform_clusterisation}{If you want to perform the second part of the analysis (clusterisation)}
}
\value{
an analysis object containing the integrated data.
}
\description{
Takes a list of samples and either a path to a cellranger directory with all the samples within, or a list of Seurat object.
}
