% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{integrate_data}
\alias{integrate_data}
\title{Integrate several analysis together.}
\usage{
integrate_data(
  analysis_list,
  method = "Seurat",
  nfeatures = 5000,
  assay = "RNA"
)
}
\arguments{
\item{analysis_list}{A list of several objects to integrate together. These objects must all be of the same type.}

\item{method}{The analysis method. Default "Seurat"}

\item{nfeatures}{The number of features to use in integration selection. Default 5000}

\item{assay}{Which assay to use. Default "RNA"}
}
\value{
An analysis object with all the analysis from the list integrated together.
}
\description{
Integrate several analysis together.
}
