% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{load_data}
\alias{load_data}
\title{Load from a tool to a R package.}
\usage{
load_data(path_to_count, sample, from = "Cellranger", to = "Seurat")
}
\arguments{
\item{path_to_count}{The path to where the directories for each sample are.
Path to the matrix inside the sample directory will be detected automatically for Cellranger.}

\item{sample}{Name of the sample. Will be added in the metadata of the returned object.}

\item{from}{Which tool does the count matrix come from. Currently only Cellranger supported.}

\item{to}{To which package should the matrix be converted to. Currently only Seurat supported.}
}
\value{
An object from the package in the argument to. By default will return a Seurat object.
}
\description{
Currently only supports from Cellranger to Seurat.
}
