% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_analysis_report}
\alias{make_analysis_report}
\title{Make the integration report, with the files to include hardcoded inside.}
\usage{
make_analysis_report(
  sample,
  report_path,
  report_name,
  plots_relative_path = "plots",
  data_relative_path = "results",
  title = "Analysis report",
  force = FALSE
)
}
\arguments{
\item{sample}{List of samples to include in the report.}

\item{report_path}{Path to where the report should be located.}

\item{report_name}{Name to give to the report file.}

\item{plots_relative_path}{Path to the plots directory, relative to the \code{report_path}.
Must be a subdirectory in \code{report_path}. Default "plots"}

\item{data_relative_path}{Path to the csv dataframes directory, relative to the \code{report_path}.
Must be a subdirectory in \code{report_path}. Default "results"}

\item{title}{The title of the report. Default to "Analysis report"}

\item{force}{Set to true if you want to overwrite the report that could already exists. Default FALSE}
}
\value{
Path to the report as a string.
}
\description{
Make the integration report, with the files to include hardcoded inside.
}
