% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_analysis_report_qmd}
\alias{make_analysis_report_qmd}
\title{Title}
\usage{
make_analysis_report_qmd(
  analysis,
  sample,
  report_path,
  file_name,
  report_steps,
  plots_relative_path = "plots",
  data_relative_path = "results",
  force = FALSE
)
}
\arguments{
\item{analysis}{}

\item{sample}{}

\item{report_path}{}

\item{file_name}{}

\item{report_steps}{}

\item{plots_relative_path}{}

\item{data_relative_path}{}

\item{force}{}
}
\description{
Title
}
