% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{make_integrate_report_qmd}
\alias{make_integrate_report_qmd}
\title{Title}
\usage{
make_integrate_report_qmd(
  analysis,
  samples,
  report_steps = report_steps,
  plots_relative_path = "plots",
  report_path = ".",
  file_name = "integrated",
  data_relative_path = "results",
  force = FALSE
)
}
\arguments{
\item{analysis}{}

\item{samples}{}

\item{report_steps}{}

\item{plots_relative_path}{}

\item{report_path}{}

\item{file_name}{}

\item{data_relative_path}{}

\item{force}{}
}
\description{
Title
}
