% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/objects.R
\name{merged}
\alias{merged}
\alias{merged.DataSheet}
\title{Combine multiple data}
\usage{
merged(object, ...)

\method{merged}{DataSheet}(object, ...)
}
\arguments{
\item{object}{an object}

\item{...}{Arguments passed to other methods}
}
\description{
Combine multiple data
}
