% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{neighbors}
\alias{neighbors}
\title{Compute neighbors}
\usage{
neighbors(analysis, method = "Seurat", k.param = 20, npcs = 50, variance = 0.8)
}
\arguments{
\item{analysis}{The analysis object.}

\item{method}{The analysis method. Default "Seurat"}

\item{k.param}{Defines k for the k-nearest neighbor algorithm. Default 20}
}
\value{
An analysis object of type method with neighbors.
}
\description{
Compute neighbors
}
