% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{normalize_data}
\alias{normalize_data}
\title{Perform the data normalization using normalize, variable features and scaling.}
\usage{
normalize_data(
  analysis,
  method = "Seurat",
  assay = "RNA",
  nfeatures = 2000,
  selection_method = "vst",
  features = NULL
)
}
\arguments{
\item{analysis}{The analysis object.}

\item{method}{The analysis method. Default "Seurat"}

\item{assay}{Which assay to make the filters on. Default "RNA"}

\item{nfeatures}{Number of variable genes to select. Default 2000}

\item{selection_method}{What method to use for features selection. Default vst}

\item{features}{Vector of features to scale. Keep NULL to scale only variable features. Default NULL}
}
\value{
An analysis object of type method normalized.
}
\description{
Perform the data normalization using normalize, variable features and scaling.
}
