% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{pca}
\alias{pca}
\title{Compute the PCA}
\usage{
pca(
  analysis,
  sample = "",
  method = "Seurat",
  assay = "RNA",
  npcs = 50,
  plots_dir = ""
)
}
\arguments{
\item{analysis}{The analysis object.}

\item{sample}{Sample name, will be used to name the images files.}

\item{method}{The analysis method. Default "Seurat"}

\item{assay}{Which assay to compute the PCA on. Default "RNA"}

\item{npcs}{Number of components to compute. Default 50}

\item{plots_dir}{The path to save the plots. Keep empty to skip plot saving. Default ""}
}
\value{
An analysis object of type method with PCA.
}
\description{
Compute the PCA
}
