% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_filter}
\alias{plot_filter}
\title{Plot qc as violin plot, adding cutoffs lines if needed.}
\usage{
plot_filter(
  df,
  x_name = "x",
  y_name = "y",
  low = 0,
  high = Inf,
  force_cutoff = FALSE
)
}
\arguments{
\item{df}{A data frame containing x_name and y_name.}

\item{x_name}{The column of df to use as groups in the violin plot, one violin per group.}

\item{y_name}{The column of df to use as y axis.}

\item{low}{The lower cutoff to put on the plot. Defaults to 0, which will not put a lower cutoff.}

\item{high}{The higher cutoff to put on the plot. Defaults to Inf, which will not put a higher cutoff.}

\item{force_cutoff}{Whether to force cutoffs to be added, even if they should not (mainly used to add a lower cutoff at 0). Defaults to FALSE.}
}
\value{
A ggplot object.
}
\description{
Plot qc as violin plot, adding cutoffs lines if needed.
}
