% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_label_umap}
\alias{plot_label_umap}
\title{Put labels on UMAPs using centroids and ggrepel.}
\usage{
plot_label_umap(
  seurat,
  assay = "RNA",
  slot = "data",
  colour_by = "orig.ident",
  label.size = 4,
  pt.size = 0.5,
  color_palette_ordered = NULL
)
}
\arguments{
\item{seurat}{}

\item{slot}{}

\item{colour_by}{}

\item{label.size}{}

\item{pt.size}{}

\item{color_palette_ordered}{}

\item{assays}{}

\item{width}{}

\item{height}{}

\item{path}{}

\item{project_name}{}
}
\description{
Put labels on UMAPs using centroids and ggrepel.
}
