% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_seurat_clustree}
\alias{plot_seurat_clustree}
\title{Plot a clustree}
\usage{
plot_seurat_clustree(seurat, prefix = "RNA_snn_res.")
}
\arguments{
\item{seurat}{The Seurat object}

\item{prefix}{The prefix to use to gather all the resolutions in the graph.
This should be the whole prefix before the resolution number. Default "RNA_snn_res."}
}
\value{
a ggplot graph
}
\description{
Clustree is a graph that shows the moving of cells between the clusters in each resolution. The clustering numbering between each resolution
}
