% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_seurat_dim}
\alias{plot_seurat_dim}
\title{Plot the cells on a reduction, with colours from the object.}
\usage{
plot_seurat_dim(
  seurat,
  reduction = "pca",
  colour_by = "orig.ident",
  assay = "RNA",
  slot = "data"
)
}
\arguments{
\item{seurat}{The Seurat object}

\item{reduction}{Which reduction to use. Default "pca"}

\item{colour_by}{Which data to use to colour cells. Default "orig.ident"}

\item{assay}{Which assay to use (if applicable). Default "RNA"}

\item{slot}{Which slot to use within the assay (if applicable). Default "data"}
}
\value{
A ggplot object
}
\description{
X axis is the first component, Y axis second component. Colour_by is fetched
from the seurat object, and can be a gene or a column from the metadata.
}
