% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_seurat_elbow}
\alias{plot_seurat_elbow}
\title{Make the Elbow plot of a dimension reduction}
\usage{
plot_seurat_elbow(seurat, reduction = "pca", npc = 50, k.param.neighbors = 20)
}
\arguments{
\item{seurat}{A Seurat object}

\item{reduction}{Which reduction to us. Must be present in the Seurat object. Default "pca"}

\item{npc}{Number of components to plot. Default 50}

\item{k.param.neighbors}{Number of k.param for neighbors joining. Default 20}
}
\value{
The elbow plot as a ggplot object
}
\description{
The elbow plot is the standard deviation in y axis and the component number in x axis.
Most of the time it has the shape of elbow, hence why it is called elbow plot.
The goal of this plot is to help choose the correct number of components for further analysis.
}
