% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_seurat_violin}
\alias{plot_seurat_violin}
\title{Make a violin plot, using the features (genes or elements of metadata) from the Seurat object.}
\usage{
plot_seurat_violin(
  seurat,
  features,
  group.by = "orig.ident",
  assay = "RNA",
  slot = "data",
  show_points = TRUE,
  taxonomy = "ensembl",
  threshold = NULL,
  alpha = 0.2,
  box_fill = "lightgrey"
)
}
\arguments{
\item{seurat}{The Seurat object from which to}

\item{show_points}{}
}
\description{
Make a violin plot, using the features (genes or elements of metadata) from the Seurat object.
}
