% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_IO.R
\name{readDB.blobs}
\alias{readDB.blobs}
\title{The API only reads data from db by chunks (entire columns)}
\usage{
readDB.blobs(db, rindx = NULL, cindx = NULL, compress = c("lz4", "gzip"))
}
\description{
The API only reads data from db by chunks (entire columns)
}
