% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{scImpute}
\alias{scImpute}
\title{scImpute algorithm}
\usage{
scImpute(dat, drop_idx, Kcluster, min_size = 5, verbose = F,
  weight = 0.5)
}
\arguments{
\item{dat}{dataset where the \code{n} rows represent cells and \code{d} columns represent genes}

\item{drop_idx}{inidices of \code{dat} to impute}

\item{Kcluster}{number of clusters}

\item{min_size}{minimum size of a cluster}

\item{verbose}{boolean}

\item{weight}{0-1 weight factor, where numbers closer to 1 means to rely on newly imputed values}
}
\value{
\code{n} by \code{d} matrix
}
\description{
scImpute algorithm
}
