% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_all_DE}
\alias{seurat_all_DE}
\title{Function to find all genetic markers per clusters}
\usage{
seurat_all_DE(
  seurat,
  sample = "",
  assay = "RNA",
  slot = "data",
  method = "wilcox",
  logfc_threshold = 0.25,
  min_pct = 0.1,
  pvalue_threshold = 0.05,
  only.pos = FALSE
)
}
\arguments{
\item{seurat}{The Seurat object.}

\item{assay}{Which assay to make the filters on. Default "RNA"}

\item{slot}{Slot to pull data from. Default "data"}

\item{method}{The statistic test that compares expression for each duo. Default "wilcox"}

\item{logfc_threshold}{The threshold for the log value. Shows result that have a difference of that value. Default 0.25}

\item{min_pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations.
Meant to spped up the function by not testing genes that are very infrequently expressed. Default 0.1}

\item{pvalue_threshold}{Threshold of significant difference between expression of markers. Default 0.05}

\item{only.pos}{Only return positive markers. Default FALSE}
}
\value{
A table with all the markers per clusters with p-values and the Fold Change log value
}
\description{
Function to find all genetic markers per clusters
}
