% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_compute_mt}
\alias{seurat_compute_mt}
\title{Get the mitochondrial percentage per cell from a list of genes in a Seurat object.}
\usage{
seurat_compute_mt(
  seurat,
  assay = "RNA",
  organism = "",
  mitochondrial_genes = NULL
)
}
\arguments{
\item{seurat}{A Seurat object}

\item{assay}{Which assay to use. Default RNA}

\item{organism}{An organism supported by the library. Currently supported : human and mouse.
"" to skip automatic mitochondrial detection. Default ""}

\item{mitochondrial_genes}{A vector of mitochondrial genes. If a supported organism is requested,
it will be skipped. Default NULL}
}
\value{
A Seurat object
}
\description{
If both organism and mitochondrial_genes are empty, no mitochondrial QC will be performed.
}
