% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_features}
\alias{seurat_features}
\title{Run Seurat::FindVariableFeatures with nfeatures and selected method}
\usage{
seurat_features(seurat, assay = "RNA", nfeatures = 2000, method = "vst")
}
\arguments{
\item{seurat}{Seurat object.}

\item{assay}{Which assay to use.}

\item{nfeatures}{Number of variable genes to select. Default 2000}

\item{method}{What to use for features selection. Default vst}
}
\value{
A Seurat object with variable features selected.
}
\description{
Run Seurat::FindVariableFeatures with nfeatures and selected method
}
