% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_filter}
\alias{seurat_filter}
\title{Apply filters on Cells and genes in a Seurat object}
\usage{
seurat_filter(
  sample,
  seurat,
  assay = "RNA",
  min_genes = 100,
  min_counts = 100,
  min_cells = 1,
  min_mt = 0,
  max_genes = Inf,
  max_counts = Inf,
  max_cells = Inf,
  max_mt = Inf,
  results_dir = ""
)
}
\arguments{
\item{sample}{Sample name, will be used to name the files.}

\item{seurat}{The seurat object.}

\item{assay}{Which assay int he seurat object to filter on. Default RNA}

\item{min_genes}{All cells having a lower number of genes expressed will be filtered out. Default 0}

\item{min_counts}{All cells having a lower number of counts will be filtered out. Default 0}

\item{min_cells}{All genes having a lower number of cells expressing it will be filtered out. Default 0}

\item{min_mt}{All cells having a lower percentage of mitochondria will be filtered out. Default 0}

\item{max_genes}{All cells having a higher number of genes expressed will be filtered out. Default Inf}

\item{max_counts}{All cells having a higher number of counts will be filtered out. Default Inf}

\item{max_cells}{All genes having a higher number of cells expressing it will be filtered out. Default Inf}

\item{max_mt}{All cells having a higher percentage of mitochondria will be filtered out. Default Inf}

\item{results_dir}{The path to save the dataframe. Keep empty to skip plot saving. Default ""}
}
\value{
A Seurat object with its count matrix filtered.
}
\description{
If no percent_mt is found in meta data, min_mt and max_mt will be skipped.
}
