% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_get_dim}
\alias{seurat_get_dim}
\title{Title}
\usage{
seurat_get_dim(seurat, reduction = "pca", npcs = 50, variance = 0.8)
}
\arguments{
\item{seurat}{}

\item{reduction}{}

\item{npcs}{}

\item{variance}{}
}
\description{
Title
}
