% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_integrate}
\alias{seurat_integrate}
\title{Integrate a list of Seurat objects together.}
\usage{
seurat_integrate(seurat_list, nfeatures = 5000, assay = "RNA")
}
\arguments{
\item{seurat_list}{A list of several objects to integrate together. These objects must all be of the same type.}

\item{nfeatures}{The number of features to use in integration selection. Default 5000}

\item{assay}{Which assay to use. Default "RNA"}
}
\value{
A Seurat object with all the analysis from the list integrated together.
}
\description{
Integrate a list of Seurat objects together.
}
