% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_load_10x}
\alias{seurat_load_10x}
\title{load Cellranger data for Seurat}
\usage{
seurat_load_10x(path_to_matrix, project = "seurat")
}
\arguments{
\item{path_to_matrix}{The path to a cellranger matrix directory}

\item{project}{The project name, put here the name of the sample currently being analyzed}
}
\value{
A Seurat object containing the data from the matrix
}
\description{
load Cellranger data for Seurat
}
