% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_neighbors}
\alias{seurat_neighbors}
\title{Compute the KNN graph using FindNeighbors from Seurat}
\usage{
seurat_neighbors(seurat, k.param = 20, dims = 20)
}
\arguments{
\item{seurat}{The Seurat object.}

\item{k.param}{Defines k for the k-nearest neighbor algorithm. Default 20}
}
\value{
A Seurat object with KNN.
}
\description{
Compute the KNN graph using FindNeighbors from Seurat
}
