% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_normalize}
\alias{seurat_normalize}
\title{Seurat NormalizeData}
\usage{
seurat_normalize(seurat, assay = "RNA")
}
\arguments{
\item{seurat}{Seurat object to use.}

\item{assay}{Assay to normalize. Default RNA}
}
\value{
A Seurat object with normalized data.
}
\description{
Seurat NormalizeData
}
