% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_pca}
\alias{seurat_pca}
\title{Compute the PCA using RunPCA from Seurat}
\usage{
seurat_pca(seurat, assay = "RNA", npcs = 50)
}
\arguments{
\item{seurat}{The Seurat object.}

\item{assay}{Assay to use. Default RNA}

\item{npcs}{Number of components to compute. Default 50}
}
\value{
A seurat object with PCA.
}
\description{
Compute the PCA using RunPCA from Seurat
}
