% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seurat.R
\name{seurat_scale}
\alias{seurat_scale}
\title{Run Seurat::ScaleData}
\usage{
seurat_scale(seurat, assay = "RNA", features = NULL)
}
\arguments{
\item{seurat}{The Seurat object.}

\item{assay}{Assay to scale. Default RNA}

\item{features}{Features to scale. Default all.genes}
}
\value{
A seurat object with scaled features.
}
\description{
Run Seurat::ScaleData
}
