% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sparse_mat.R
\name{simulate_sparse_mat}
\alias{simulate_sparse_mat}
\title{simulate single cell data as the sparse matrix
This is the slow version is deprecated by simulate_sparse_mat_fast}
\usage{
simulate_sparse_mat(alpha, beta, nCells)
}
\description{
simulate single cell data as the sparse matrix
This is the slow version is deprecated by simulate_sparse_mat_fast
}
