% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sparse_mat.R
\name{simulate_sparse_mat_v3}
\alias{simulate_sparse_mat_v3}
\title{more efficient version (gene x cells) by preallocating vector}
\usage{
simulate_sparse_mat_v3(alpha, beta, nCells, cpp = FALSE)
}
\description{
more efficient version (gene x cells) by preallocating vector
}
