% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split.block}
\alias{split.block}
\title{split the vector into groups so that the disk IO will be performed block-wise}
\usage{
\method{split}{block}(x, block.size = 1000)
}
\arguments{
\item{x}{a vector to be splitted}
}
\description{
split the vector into groups so that the disk IO will be performed block-wise
}
