% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{umap}
\alias{umap}
\title{Compute and plot the umap.}
\usage{
umap(
  analysis,
  sample = "",
  method = "Seurat",
  n_neighbors = 30,
  plot_clustering = "RNA_snn_res.1",
  plots_dir = "",
  npcs = 50,
  variance = variance
)
}
\arguments{
\item{analysis}{The analysis object.}

\item{sample}{Sample name, will be used to name the images files.}

\item{method}{The analysis method. Default "Seurat"}

\item{n_neighbors}{number of neighbors to use when computing the UMAP. Default 30}

\item{plot_clustering}{Which clustering to colour the UMAP with. Default "RNA_snn_res.1"}

\item{plots_dir}{The path to save the plots. Keep empty to skip plot saving. Default ""}
}
\value{
an analysis object.
}
\description{
Will plot a first umap showing the orig.ident,
a second one using the column in plot_clustering.
}
