% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiledb_IO.R
\name{write_tiledb_dense}
\alias{write_tiledb_dense}
\title{write tile db as sparse matrix from dense matrix}
\usage{
write_tiledb_dense(mat, tiledb_dir, tiledb_attr)
}
\description{
write tile db as sparse matrix from dense matrix
}
