% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql_IO.R
\name{writeDB.blobs}
\alias{writeDB.blobs}
\title{Create sqlite db from matrix
Each column is stored as binary blobs and indexed by column id}
\usage{
writeDB.blobs(mat, dbfile, nGenes, compress = c("lz4", "gzip"))
}
\description{
Create sqlite db from matrix
Each column is stored as binary blobs and indexed by column id
}
