% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdb_IO.R
\name{writelmdb.blobs}
\alias{writelmdb.blobs}
\title{Create sqlite db from matrix
Each column is stored as binary blobs and indexed by column id}
\usage{
writelmdb.blobs(mat, dbfile, nGenes, compress = c("lz4", "gzip", "none"))
}
\description{
Create sqlite db from matrix
Each column is stored as binary blobs and indexed by column id
}
