% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{analyze_integrated}
\alias{analyze_integrated}
\title{Title}
\usage{
analyze_integrated(
  analysis,
  sample = "integrated",
  step = 6,
  method = "Seurat",
  assay = "RNA",
  save_path = ".",
  file_name = "analysis.rds",
  organism = "",
  mitochondrial_genes = c(),
  min_genes = 100,
  min_counts = 100,
  min_cells = 1,
  min_mt = 0,
  max_genes = Inf,
  max_counts = Inf,
  max_cells = Inf,
  max_mt = 20,
  nfeatures_normalize = 2000,
  selection_method_normalize = "vst",
  npcs_pca = 50,
  k.param_neighbors = 20,
  resolutions_clustree = c(1:10/10, 5:8/4),
  resolution_clustering = 1,
  n_neighbors = 30,
  de_test = "wilcox",
  de_logfc = 0.25,
  de_pvalue = 0.05,
  force_report = FALSE
)
}
\arguments{
\item{analysis}{Object to use.}

\item{sample}{Sample name.}

\item{step}{Steps range from 6 to 11, following the steps from previous function \code{integrate},
with each number making the function start at one of the following steps :
6 : \verb{filter data}, 7 : \verb{normalize data}, 8 : \verb{compute pca}, 9 : \verb{find neighbors}, 10 : \verb{find clusters}, 11 : \verb{compute UMAP}.
Choosing 6 will do the entire process, while 5 will only do the umap and assume that every previous step needed has already been done by the user.
Default 6}

\item{method}{The analysis method used through the analysis, as well as the objects type in \code{analysis_list}.
Currently supports : Seurat. Default "Seurat"}

\item{assay}{Which to use for all functions asking for it. Default "RNA"}

\item{save_path}{In which directory the final analysis is saved, in rds format.
Keep empty to just return the analysis and not save it. Default "."}

\item{file_name}{What name to give to the final file. Will be skipped if \code{save_path} is empty. Default "analysis.rds"}

\item{organism}{The organism is used for mitochondrial genes detection. Supports mouse and human.
Keep empty to skip mitochondrial detection. Default ""}

\item{mitochondrial_genes}{List of mitochondrial genes. This argument will be completely ignored}

\item{min_genes}{All cells having a lower number of genes expressed will be filtered out. Default 0}

\item{min_counts}{All cells having a lower number of counts will be filtered out. Default 0}

\item{min_cells}{All genes having a lower number of cells expressing it will be filtered out. Default 0}

\item{min_mt}{All cells having a lower percentage of mitochondria will be filtered out. Default 0}

\item{max_genes}{All cells having a higher number of genes expressed will be filtered out. Default Inf}

\item{max_counts}{All cells having a higher number of counts will be filtered out. Default Inf}

\item{max_cells}{All genes having a higher number of cells expressing it will be filtered out. Default Inf}

\item{max_mt}{All cells having a higher percentage of mitochondria will be filtered out. Default Inf}

\item{nfeatures_normalize}{Number of variable genes to select for scaling. Default 2000}

\item{selection_method_normalize}{What method to use for features selection during scaling. Default "vst"}

\item{npcs_pca}{Number of principal components to compute in the PCA. Default 50}

\item{k.param_neighbors}{Number of neighbors to use when computing UMAP. Default 30}

\item{resolutions_clustree}{Which resolutions to compute and run the clustree on.
If this parameter is NULL, clustree will be skipped. Clustree is an optional part of step 10. Default c(1:10/10, 5:8/4)}

\item{resolution_clustering}{Which final resolution to keep (can be a resolution not in \code{resolutions_clustree}).
Typically resolutions range between 0.1 and 2. Default 1}

\item{n_neighbors}{\subsection{}{
}}

\item{de_test}{Which statistical tests to use for DE. To see available options, see the documentation for Seurat::FindAllMarkers. Default "wilcox"}

\item{de_logfc}{Filter genes based on a minimum logged fold change. Default 0.25}

\item{de_pvalue}{Filter genes based on a maximum pvalue. Default 0.05}

\item{force_report}{Whether to overwrite the report if it already exists. If this is FALSE, and a file \code{analysis.Rmd} already exists,
report creation will be skipped. Default FALSE}
}
\value{
An analysis object.
}
\description{
Title
}
