% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slingshot_lineages.R
\name{.get_lineages}
\alias{.get_lineages}
\title{Estimate the lineages (via Slingshot)}
\usage{
.get_lineages(dat, cluster_labels, starting_cluster,
  cluster_group_list = NA)
}
\arguments{
\item{dat}{a \code{n} by \code{d} matrix}

\item{cluster_labels}{vector of cluster labels, where
the cluster labels are consecutive positive integers from 1 to
\code{max(cluster_labels)}. Cannot include \code{NA}}

\item{starting_cluster}{the "origin" cluster that all the lineages will start
from}

\item{cluster_group_list}{list denoting the hierarchy and order of the clusters}
}
\value{
A list of cluster indices, with \code{starting_cluster} starting as
its first element
}
\description{
Note: I removed the functionality to explicitly label a ending cluster
(might put back in later?).
}
\details{
Note: I removed the Omega parameter, which, to my understanding,
controls if a lineage (i.e. tree) is split into two separate lineages.

Code adapted from https://github.com/kstreet13/slingshot.
}
