% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_factorization.R
\name{.initialization}
\alias{.initialization}
\title{Initialization function}
\usage{
.initialization(dat, k = 2)
}
\arguments{
\item{dat}{matrix}

\item{k}{numeric}
}
\value{
list
}
\description{
Initialization function
}
