% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{.projection_l1}
\alias{.projection_l1}
\title{L1 projection of the product of two vectors to become nonpositive}
\usage{
.projection_l1(current_vec, other_mat, tol = 0.001, direction = "<=",
  other_bound = NA)
}
\arguments{
\item{current_vec}{vector}

\item{other_mat}{matrix}

\item{tol}{numeric}

\item{direction}{character}

\item{other_bound}{numeric}
}
\description{
Solves the linear program:
Data: u_i (vector), i from 1 to k. V (matrix), with d rows and k columns.
Variables: s_1 to s_k, z_1 to z_k
Optimization problem: min s_1 + ... + s_k
such that: s_i >= u_i - z_i for i from 1 to k,
           s_i >= -(u_i - z_i) for i from 1 to k,
           V %*% z_k <= tol elementwise (for entire vector of length d)
           s_i >= 0 for i from 1 to k
}
