% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{normalize_data}
\alias{normalize_data}
\title{Perform the data normalization using normalize, variable features and scaling.}
\usage{
normalize_data(
  analysis,
  method = "Seurat",
  assay = "RNA",
  nfeatures = 2000,
  selection_method = "vst",
  features = all.genes
)
}
\arguments{
\item{features}{}
}
\description{
Perform the data normalization using normalize, variable features and scaling.
}
