% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_filtering_stats}
\alias{plot_filtering_stats}
\title{Barplots that represent the filtering statistics for the seurat filtration}
\usage{
plot_filtering_stats(df)
}
\arguments{
\item{df}{dataframe with the statistics}
}
\value{
Barplots as a ggplot object
}
\description{
Barplots that represent the filtering statistics for the seurat filtration
}
