% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdb_IO.R
\name{readlmdb.blobs}
\alias{readlmdb.blobs}
\title{The API only reads data from db by chunks (entire columns)}
\usage{
readlmdb.blobs(dbfile, rindx = NULL, cindx = NULL, compress = c("lz4",
  "gzip"))
}
\description{
The API only reads data from db by chunks (entire columns)
}
